+++ packages/dynamic-flows/src/jsonSchemaForm/basicTypeSchema/BasicTypeSchema.spec.js
@@ -111,6 +111,25 @@ describe('Given a component for rendering basic type schemas', () => {
       it('should tell the ControlFeedback it has been changed', () => {
         expect(feedbackComponent.prop('changed')).toEqual(true);
       });
+
+      it('should broadcast null value for optional field when valid input changes to empty string', () => {
+        model = 'foo';
+        const isRequired = false;
+        props = { schema, model, errors, isRequired, onChange, submitted, translations };
+        component = shallow(<BasicTypeSchema {...props} />);
+
+        formControl = component.find(SchemaFormControl);
+        feedbackComponent = component.find(ControlFeedback);
+
+        // valid non null value
+        formControl.simulate('change', 'barbar');
+        expect(onChange).toHaveBeenCalledWith('barbar', schema);
+
+        // empty value but is still valid
+        formControl.simulate('change', '');
+        expect(onChange).toHaveBeenCalledWith(null, schema);
+        expect(feedbackComponent.prop('validations').length).toEqual(0);
+      });
     });
 
     describe('when control triggers onChange with an invalid value', () => {
@@ -148,8 +167,8 @@ describe('Given a component for rendering basic type schemas', () => {
         expect(formControl.prop('value')).not.toEqual(schema.default);
       });
 
-      it('should pass empty value to the form control', () => {
-        expect(formControl.prop('value')).toBe('');
+      it('should pass null value to the form control', () => {
+        expect(formControl.prop('value')).toBe(null);
       });
     });
   });