+++ packages/dynamic-flows/src/jsonSchemaForm/schemaFormControl/SchemaFormControl.js
@@ -4,6 +4,7 @@ import Types from 'prop-types';
 import FormControl from '../../formControl';
 import { isNull, isUndefined } from '@transferwise/neptune-validation';
 import { getValidModelParts } from '../../common/validation/valid-model';
+import { availableCurrencyFlags } from './availableCurrencyFlags';
 
 const SchemaFormControl = (props) => {
   const isNativeInput = (schemaType) => schemaType === 'string' || schemaType === 'number';
@@ -12,7 +13,7 @@ const SchemaFormControl = (props) => {
     isNativeInput(props.schema.type) && (isNull(value) || isUndefined(value)) ? '' : value;
 
   const onChange = (value) => {
-    // If the model does not satisfy the schema propogate null
+    // If the model does not satisfy the schema propagate null
     props.onChange(getValidModelParts(value, props.schema));
   };
 
@@ -67,11 +68,25 @@ const SchemaFormControl = (props) => {
     return null;
   };
 
+  const mapIcon = (icon) => {
+    if (icon) {
+      if (icon.name && availableCurrencyFlags.includes(icon.name)) {
+        return {
+          currency: icon.name,
+        };
+      }
+    }
+    return null;
+  };
+
   const mapConstSchemaToOption = (schema) => {
+    const keyForDescription =
+      (schema.title + schema.description).length > 50 ? 'secondary' : 'note';
     return {
       value: !isUndefined(schema.const) ? schema.const : schema.enum[0],
       label: schema.title,
-      secondary: schema.description,
+      [keyForDescription]: schema.description,
+      ...mapIcon(schema.icon),
       disabled: schema.disabled,
     };
   };