+++ packages/dynamic-flows/src/jsonSchemaForm/schemaFormControl/SchemaFormControl.spec.js
@@ -202,6 +202,78 @@ describe('Given a component for rendering a form control based on a schema', () 
         { value: 3, label: 'Three' },
       ]);
     });
+
+    it('should set secondary field if total length of label and description exceeds predefined length', () => {
+      component.setProps({
+        schema: {
+          oneOf: [
+            { const: 1, title: 'One' },
+            {
+              const: 3,
+              title: 'Three',
+              description: 'reeeeeeaaallllllllly loooooooooooooooooooooongggggggggggg',
+            },
+          ],
+        },
+      });
+
+      expect(component.find(FormControl).prop('options')).toEqual([
+        { value: 1, label: 'One' },
+        {
+          value: 3,
+          label: 'Three',
+          secondary: 'reeeeeeaaallllllllly loooooooooooooooooooooongggggggggggg',
+        },
+      ]);
+    });
+
+    it('should set note field if total length of label and description is less than predefined length', () => {
+      component.setProps({
+        schema: {
+          oneOf: [
+            { const: 1, title: 'One' },
+            { const: 3, title: 'Three', description: 'short description' },
+          ],
+        },
+      });
+
+      expect(component.find(FormControl).prop('options')).toEqual([
+        { value: 1, label: 'One' },
+        { value: 3, label: 'Three', note: 'short description' },
+      ]);
+    });
+
+    it('should set currency field when icon name is a supported currency flag', () => {
+      component.setProps({
+        schema: {
+          oneOf: [
+            { const: 1, title: 'One' },
+            { const: 3, title: 'Three', icon: { name: 'gbp' } },
+          ],
+        },
+      });
+
+      expect(component.find(FormControl).prop('options')).toEqual([
+        { value: 1, label: 'One' },
+        { value: 3, label: 'Three', currency: 'gbp' },
+      ]);
+    });
+
+    it('should not set currency field when icon name is not a supported currency flag', () => {
+      component.setProps({
+        schema: {
+          oneOf: [
+            { const: 1, title: 'One' },
+            { const: 3, title: 'Three', icon: { name: 'something invalid' } },
+          ],
+        },
+      });
+
+      expect(component.find(FormControl).prop('options')).toEqual([
+        { value: 1, label: 'One' },
+        { value: 3, label: 'Three' },
+      ]);
+    });
   });
 
   describe('when a schema has a control value', () => {