+++ packages/dynamic-flows/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.6.0](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@3.5.0...@transferwise/dynamic-flows@3.6.0) (2020-08-28)
+
+
+### Features
+
+* ADS-1539 empty validation on optional field ([#541](https://github.com/transferwise/neptune-web/issues/541)) ([551c6f3](https://github.com/transferwise/neptune-web/commit/551c6f3253f7d4873ae1e6cc619302448be5ea21))
+
+
+
+
+
+# [3.5.0](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@3.4.1...@transferwise/dynamic-flows@3.5.0) (2020-08-28)
+
+
+### Features
+
+* ADS-1539 currency selects ([#539](https://github.com/transferwise/neptune-web/issues/539)) ([9c82dad](https://github.com/transferwise/neptune-web/commit/9c82dadcc28ae751dc311fc4a4bd7fda29b18763))
+
+
+
+
+
 ## [3.4.1](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@3.4.0...@transferwise/dynamic-flows@3.4.1) (2020-08-19)
 
 **Note:** Version bump only for package @transferwise/dynamic-flows